fixed the default fruits array bug for every level and optimized components
