+++ src/zustand/useLevel.js
@@ -3,7 +3,7 @@ import { create } from 'zustand'
 const useLevel = create((set) => ({
     level: 1,
     setLevel: (level) => set({ level }),
-    fruits: ['🍎', '🍌', '🍇', '🍓', '🍍'],
+    fruits: [],
     setFruits: (fruits) => set({ fruits })
 }))
